Plastic types at beach and litterboom

Weight in kg of identified PET, PE, PP, PS, other plastic and other waste items.

Usage

plastic_types

Format

A tibble with 13 rows and 16 variables:#'

date

Collection date of the data

place

Place where the collection was done

total

Total amount of waste chategorized in kg.

pet

Amount of polyethylene terephthalate (PET) in kg.

pe

Amount of polyethylene (PE) in kg. During the sorting process HDPE and LDPE plastics were not distinguished.

pp

Amount of polypropylene (PP) in kg.

ps

Amount of polystyrene (PS) in kg.

hdpe_pp

Amount of high-density polyethylene (HDPE) and polypropylene (PP) combined weight in kg. During the sorting process these two plastics were not distinguished.

other_plastics

Amount of other plastics in kg. The category ‘Other plastics’ included plastics that couldn’t be identified by the scanner or plastics that are in such small amounts the weighing scale could not distinguish it.

other_waste

Amount of other waste in kg remaining at the end of the categorization.

glass

Amount of glass in kg.

shoes

Amount of shoes in kg.

shoes_quantity

Number of shoes found at the location. Surprisingly a lot of shoes were found. As they often consisted of a combination of different plastics then a separate category was made for shoes.

bag_quantity

Number of bags that where collected.

beach

Categorical variable showing whether the collection was done at the beach (TRUE) or at the a litterboom (FALSE).

details

Additional information about the sampled data.